Moved App.tsx into /component
