+++ src/components/App.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
-import "./CSS/App.css";
-import HomeScreen from "./components/HomeScreen";
-import AppHeader from "./components/AppHeader";
-import Footer from "./components/Footer";
-import Gallery from "./components/Gallery";
-import PageNotFound from "./components/PageNotFound";
-import ScrollToTop from "./components/ScrollToTop";
-import { HOME, GALLERY } from "./globals/Paths";
+import "../CSS/App.css";
+import HomeScreen from "./HomeScreen";
+import AppHeader from "./AppHeader";
+import Footer from "./Footer";
+import Gallery from "./Gallery";
+import PageNotFound from "./PageNotFound";
+import ScrollToTop from "./ScrollToTop";
+import { HOME, GALLERY } from "../globals/Paths";
 
 function App() {
     return (
@@ -19,10 +19,12 @@ function App() {
             <BrowserRouter>
                 <ScrollToTop />
                 <AppHeader />
+
                 <div className="ContentWrap">
                     <Switch>
                         <Route component={HomeScreen} exact path={HOME} />
                         <Route component={Gallery} exact path={GALLERY} />
+
                         <Redirect exact from="/" to={HOME} />
                         <Route path="*" component={PageNotFound} />
                     </Switch>
